pointer debug and nil supported

+++ libft/ft_ultoabase.c
@@ -6,7 +6,7 @@
 /*   By: vlageard <vlageard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/11 15:41:38 by vlageard          #+#    #+#             */
-/*   Updated: 2020/01/11 15:45:41 by vlageard         ###   ########.fr       */
+/*   Updated: 2020/01/13 19:13:16 by vlageard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,12 +34,12 @@ static int		is_base_valid(const char *base)
 	return (1);
 }
 
-static int		get_str_size(unsigned int n, int base_len)
+static int		get_str_size(unsigned long n, unsigned long base_len)
 {
 	int	str_size;
 	
 	str_size = 0;
-	while (n >= (unsigned long)base_len)
+	while (n >= base_len)
 	{
 		str_size++;
 		n = n / base_len;
@@ -50,19 +50,19 @@ static int		get_str_size(unsigned int n, int base_len)
 
 char			*ft_ultoabase(unsigned long n, const char *base)
 {
-	char *str;
-	int	base_len;
-	int	str_size;
-	int	i;
+	char 			*str;
+	unsigned long	base_len;
+	int				str_size;
+	int				i;
 	
 	if (!is_base_valid(base))
 		return (NULL);
-	base_len = ft_strlen(base);
+	base_len = (unsigned long)ft_strlen(base);
 	str_size = get_str_size(n, base_len);
 	if (!(str = (char *)malloc(sizeof(char) * (str_size + 1))))
 		return (NULL);
 	i = str_size - 1;
-	while (n >= (unsigned long)base_len)
+	while (n >= base_len)
 	{
 		str[i] = base[n % base_len];
 		n = n / base_len;
